+++ CHANGELOG.md
@@ -834,7 +834,7 @@ Promptbook server has OpenAI compatibility
 -   You can call `book` personas as any other OpenAI model
 -   Alongside `OpenAiExecutionTools` and `OpenAiAssistantExecutionTools` add `OpenAiCompatibleExecutionTools` as registration of configuration and the constructor
 -   Logging all failed results not just last result
--   Do not cache failed results and bring `DEFAULT_MAX_EXECUTION_ATTEMPTS` down to 3
+-   Do not cache failed results and bring `DEFAULT_MAX_EXECUTION_ATTEMPTS` down to 7
 
 ## Upcoming